+++ src/main/java/fun/pancakes/planet_pancakes/service/resource/PriceCalculator.java
@@ -15,31 +15,31 @@ public class PriceCalculator {
         this.random = random;
     }
 
-    public Long determineResourceNewPrice(Long currentPrice, Integer priceTrendPercent, Integer maxChangePercent) {
+    public Long determineResourceNewPrice(Long currentPrice, Double priceTrendPercent, Double maxChangePercent) {
         Long priceChange = priceChangeAmount(currentPrice, priceTrendPercent, maxChangePercent);
         long newPrice = currentPrice + priceChange;
         return newPrice < 1 ? 1 : newPrice;
     }
 
-    private Long priceChangeAmount(Long currentPrice, Integer priceTrendPercent, Integer maxChangePercent) {
+    private Long priceChangeAmount(Long currentPrice, Double priceTrendPercent, Double maxChangePercent) {
         Long randomPriceChange = randomPriceChangeAmount(currentPrice, maxChangePercent);
         Long trendPriceChange = trendPriceChangeAmount(currentPrice, priceTrendPercent);
         return randomPriceChange + trendPriceChange;
     }
 
-    private Long randomPriceChangeAmount(Long currentPrice, Integer maxChangePercent) {
+    private Long randomPriceChangeAmount(Long currentPrice, Double maxChangePercent) {
         Double zeroCenteredRandomNumber = random.nextDouble() * 2 - 1;
         Double maxChangeDecimal = decimalFromPercentage(maxChangePercent);
         Double randomChange = maxChangeDecimal * zeroCenteredRandomNumber;
         return Math.round(currentPrice * randomChange);
     }
 
-    private Long trendPriceChangeAmount(Long currentPrice, Integer priceTrendPercent) {
+    private Long trendPriceChangeAmount(Long currentPrice, Double priceTrendPercent) {
         Double priceChangeDecimal = decimalFromPercentage(priceTrendPercent);
         return Math.round(currentPrice * priceChangeDecimal);
     }
 
-    private Double decimalFromPercentage(Integer percentage) {
+    private Double decimalFromPercentage(Double percentage) {
         return percentage / 100d;
     }
